Floats are more than 3x faster than mixed floats and doubles!

+++ cuda/inOneWeekend/objects/hittable.cuh
@@ -10,7 +10,7 @@
 
 class hittable {
   public:
-    __device__ virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
+    __device__ virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
     __device__ virtual int getObjectNumber() {
       return 0;
     }
